#240 Move classes to separate files
